GA_COOKIE_AGE must be a number

+++ upkook_web/settings/web_base.py
@@ -50,7 +50,7 @@ PURE_CHAT_ID = get_env_var('PURE_CHAT_ID')
 GA_VERSION = get_env_var('GA_VERSION', 'GA1')
 
 GA_COOKIE_NAME = get_env_var('GA_COOKIE_NAME', '_ga')
-GA_COOKIE_AGE = get_env_var('GA_COOKIE_AGE', 2 * 365 * 24 * 60 * 60)  # 2 years
+GA_COOKIE_AGE = int(get_env_var('GA_COOKIE_AGE', 2 * 365 * 24 * 60 * 60))  # 2 years
 GA_COOKIE_PATH = get_env_var('GA_COOKIE_PATH', '/')
 GA_COOKIE_DOMAIN = get_env_var('GA_COOKIE_DOMAIN')
 GA_COOKIE_SECURE = get_env_var('GA_COOKIE_SECURE', 'False') == 'True'
